requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
cryptography==41.0.7
Django==4.2.7
django-appconf==1.0.6
django-debug-toolbar==4.2.0
django-ipware==6.0.4
django-js-asset==2.2.0
django-mptt==0.15.0
django-resized==1.0.2
djangorestframework==3.14.0
drf-extra-fields==3.7.0
filetype==1.2.0
idna==3.4
image==1.5.33
mysql-connector-python==8.2.0
Pillow==10.1.0
protobuf==4.21.12
pycparser==2.21
PyMySQL==1.1.0
python-dateutil==2.8.2
python-ipware==2.0.1
pytils==0.4.1
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.7
uuid==1.30
webptools==0.0.9
whitenoise==6.6.0
